requirements.txt
discord.py>=2.4.0
googletrans==4.0.0-rc1
